backend/requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
air-benchmark==0.1.0
annotated-types==0.7.0
anyio==4.7.0
argcomplete==1.10.3
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.8.2
beir==2.0.0
bitsandbytes==0.45.0
bleach==6.2.0
boto3==1.35.90
botocore==1.35.90
cbor==1.0.0
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
cmake==3.26.4
compressed_rtf==1.0.6
dataclasses-json==0.6.7
datasets==2.19.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
elasticsearch==7.9.1
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
extract-msg==0.28.7
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.12.2
filetype==1.2.0
FlagEmbedding @ git+https://github.com/FlagOpen/FlagEmbedding.git@ff3c36a6661242f1e94b1d578f4e92d950fe33b0
frozenlist==1.5.0
fsspec==2024.3.1
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.24.0
idna==3.4
ijson==3.3.0
IMAPClient==2.1.0
importlib_metadata==8.5.0
inscriptis==2.5.0
ipython==8.12.3
ir_datasets==0.5.9
jedi==0.19.2
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.3.12
langchain-core==0.3.25
langchain-text-splitters==0.3.3
langsmith==0.2.3
lit==16.0.6
llama-cloud==0.1.7
llama-index==0.12.8
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.9
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-huggingface==0.4.1
llama-index-llms-openai==0.3.12
llama-index-multi-modal-llms-openai==0.4.1
llama-index-postprocessor-flag-embedding-reranker==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-milvus==0.5.0
llama-parse==0.5.19
loguru==0.7.3
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.23.2
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.11
minijinja==2.5.0
mistune==3.1.0
mpmath==1.3.0
mteb==1.25.9
multidict==6.1.0
multiprocess==0.70.16
multithreading==0.2.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numpy==1.25.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
openai==1.58.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20191110
peft==0.13.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.5.0
platformdirs==4.3.6
polars==1.18.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.6
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pymilvus==2.5.2
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==0.6.23
pytrec-eval-terrier==0.5.6
pytrec_eval==0.5
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.13.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
striprtf==0.0.26
sympy==1.13.1
tenacity==9.0.0
text-generation==0.7.0
textract==1.6.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.19.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.2
trec-car-tools==2.6
triton==3.1.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
unlzw3==0.2.3
urllib3==1.26.20
uuid==1.30
uvicorn==0.34.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.0
xlrd==1.2.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
zlib-state==0.1.9
